+++ app/Http/Controllers/TagsController.php
@@ -92,6 +92,10 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        if($tag->posts->count() > 0){
+            session()->flash('error','Tag cannot be deleted because it is linked to some posts.');
+            return redirect()->back();
+        }
         $tag->delete();
         session()->flash('success','Tag Deleted Successfully');
         return redirect(route('tags.index'));
